Add the authorization concern to authorize access tokens
